requirements.txt
enum34; python_version < "3.4"
attrs
jinja2
decorator
pyrsistent
qualname
six
slotted
typing; python_version < "3.5"